Backend/requirements.txt
Flask
flask-cors
python-dotenv
openai
psycopg2-binary
SQLAlchemy
flask_sqlalchemy